fix: edit and delete condition added (if it has reservations it can't be deleted or edited)

+++ IsaClient/src/app/components/boat-components/boat-preview/boat-preview.component.ts
@@ -1,5 +1,6 @@
 import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
 import { MatDialog, MatDialogConfig } from '@angular/material/dialog';
+import { MatSnackBar } from '@angular/material/snack-bar';
 import { Router } from '@angular/router';
 import { Boat } from 'src/app/interfaces/boat';
 import { BoatService } from 'src/app/services/boat-service/boat.service';
@@ -24,7 +25,7 @@ export class BoatPreviewComponent implements OnInit {
   baseUrl = environment.apiURL
 
   constructor(private _router : Router, private _matDialog : MatDialog,
-              private _boatService : BoatService, public storage : StorageService) { }
+              private _boatService : BoatService, public storage : StorageService,private _snackBar : MatSnackBar) { }
 
   ngOnInit(): void {
   }
@@ -34,31 +35,52 @@ export class BoatPreviewComponent implements OnInit {
   }
 
   edit(){
+    this._boatService.incomingReservationExists(this.boat.id).subscribe(res => {
+      if (res) {
+        this._snackBar.open('This boat has booked reservations. You cannot edit it.', '', {
+          duration: 2000,
+          verticalPosition: 'top',
+          panelClass: 'snack-bar'
+        });
+      } else {
     
-    let myData = {
-      editMode : true,
-      boatId  : this.boat.id
-    }
-    const dialogConfig = new MatDialogConfig();
-    dialogConfig.disableClose = false;
-    dialogConfig.id = 'modal-component';
-    dialogConfig.width = '1100px';
-    dialogConfig.data = myData;
-    const dialogRef = this._matDialog.open(AddBoatComponent, dialogConfig);
+      let myData = {
+        editMode : true,
+        boatId  : this.boat.id
+      }
+      const dialogConfig = new MatDialogConfig();
+      dialogConfig.disableClose = false;
+      dialogConfig.id = 'modal-component';
+      dialogConfig.width = '1100px';
+      dialogConfig.data = myData;
+      const dialogRef = this._matDialog.open(AddBoatComponent, dialogConfig);
 
-    dialogRef.afterClosed().subscribe(res =>
-      {
-        if(res.data.editMode){
-        this.boatEdited.emit(res.data.boat);
-        }
-        
-      })
+      dialogRef.afterClosed().subscribe(res =>
+        {
+          if(res.data.editMode){
+          this.boatEdited.emit(res.data.boat);
+          }
+          
+        })
+      }
+    });
   }
 
   delete(){
-    this._boatService.deleteBoat(this.boat.id).subscribe(() => {
-      this.boatDeleted.emit(this.boat.id);
+    this._boatService.incomingReservationExists(this.boat.id).subscribe(res => {
+      if (res) {
+        this._snackBar.open('This boat has booked reservations. You cannot delete it.', '', {
+          duration: 2000,
+          verticalPosition: 'top',
+          panelClass: 'snack-bar'
+        });
+      } else {
+        this._boatService.deleteBoat(this.boat.id).subscribe(() => {
+          this.boatDeleted.emit(this.boat.id);
+        });
+      }
     });
+    
   }
 
   openCalendar(){
